requirements.txt
bencodepy
torch
transformers
aria2[tui]